modularized deck server functiions from routes
